+++ CommHandler/src/main/java/com/smartgridready/communicator/common/api/values/BitmapValue.java
@@ -2,6 +2,9 @@ package com.smartgridready.communicator.common.api.values;
 
 import com.smartgridready.ns.v0.BitmapEntryProduct;
 import com.smartgridready.ns.v0.BitmapProduct;
+import com.fasterxml.jackson.databind.JsonNode;
+import com.fasterxml.jackson.databind.node.JsonNodeFactory;
+import com.fasterxml.jackson.databind.node.ObjectNode;
 import com.smartgridready.communicator.modbus.helper.ConversionHelper;
 import org.apache.commons.lang3.builder.EqualsBuilder;
 import org.apache.commons.lang3.builder.HashCodeBuilder;
@@ -106,13 +109,20 @@ public class BitmapValue extends Value {
     @Override
     public Map<String, Boolean> getBitmap() {
         Map<String, Boolean> res = new LinkedHashMap<>();
-        value.forEach( bitmapRecord -> res.put(bitmapRecord.getLiteral(), bitmapRecord.getValue()));
+        value.forEach(bitmapRecord -> res.put(bitmapRecord.getLiteral(), bitmapRecord.getValue()));
         return res;
     }
 
     @Override
     public Instant getDateTime() {
         throw new UnsupportedOperationException("Cannot convert bitmap to datetime.");
+    }
+
+    @Override
+    public JsonNode getJson() {
+        ObjectNode res = new ObjectNode(JsonNodeFactory.instance);
+        value.forEach(bitmapRecord -> res.put(bitmapRecord.getLiteral(), bitmapRecord.getValue()));
+        return res;
     }
 
     @Override