+++ CommHandler/src/main/java/com/smartgridready/communicator/common/api/values/BooleanValue.java
@@ -3,6 +3,9 @@ package com.smartgridready.communicator.common.api.values;
 import org.apache.commons.lang3.builder.EqualsBuilder;
 import org.apache.commons.lang3.builder.HashCodeBuilder;
 
+import com.fasterxml.jackson.databind.JsonNode;
+import com.fasterxml.jackson.databind.node.BooleanNode;
+
 import java.math.BigInteger;
 import java.time.Instant;
 import java.util.Map;
@@ -88,6 +91,11 @@ public class BooleanValue extends Value {
     @Override
     public Instant getDateTime() {
         return Instant.ofEpochMilli(getInt64());
+    }
+
+    @Override
+    public JsonNode getJson() {
+        return BooleanNode.valueOf(value);
     }
 
     @Override